requirements.txt
psycopg[binary]==3.1.9
psycopg_pool==3.1.7
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0